#4549: Fixed compatibility between savegames of Linux and Windows builds.
1. All EventDef-s are sorted alphabetically by name (which is unique) when idEvent is initialized.
   This ensures deterministic and platform independent order of EventDefs, and consequentally, script programs.
2. Fixed memory layout of TimerValue structure on 64-bit Linux (long is 64-bit).
   Also added static assert to make sure that the layout is as expected.
   And changed save/restore functions to work via two ints instead of raw bytes.



git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7000 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/StimResponse/StimResponseTimer.h
@@ -39,7 +39,7 @@ public:
 		signed short Millisecond;
 	} Time;
 	struct {
-        signed long TimerVal;
+		signed int TimerVal;
 		signed short Millisecond;
 	} Val;
 } TimerValue;
